Add fileStorageApiMock in test utils
